Add InTheHand.BluetoothLE to BluetoothTestFramework (error).
Unity has problems loading this, producing exception.

Assembly 'Assets/Packages/InTheHand.BluetoothLE.4.0.37/lib/net462/InTheHand.BluetoothLE.dll' will not be loaded due to errors:
Unable to resolve reference 'Windows.Foundation.UniversalApiContract'. Is the assembly missing or incompatible with the current platform?
Reference validation can be disabled in the Plugin Inspector.
Unable to resolve reference 'Windows.Foundation.FoundationContract'. Is the assembly missing or incompatible with the current platform?
Reference validation can be disabled in the Plugin Inspector.
Unable to resolve reference 'System.Runtime.WindowsRuntime'. Is the assembly missing or incompatible with the current platform?
Reference validation can be disabled in the Plugin Inspector.